feat: Improve the sender name retrieval in NotificationSerializer
Refactored the get_sender_name method in the NotificationSerializer to properly handle cases where the sender's name is not set in the user profile. Now, if the sender has set their name, it is returned; otherwise, their email is used as the default sender name.
Refactored the send_notification method to include the workspace ID
Started the creation of the notification dropdown in PrivateNavbar.tsx